Implement stable last-modified times
When checking out from git, it sets the last-modified to the current time.
This causes churn in rendered files, as the mtime on the github actions
runner basically reflects when the job was run, not when the file actually
changed.

This change uses `git-restore-mtime` to restore mtimes from git commits
changing the file, thus providing a more accurate reflection of when
content has changed.